requirements.txt
# Core dependencies
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Voice processing
SpeechRecognition==3.10.0
pyttsx3==2.90

# HTTP client
aiohttp==3.9.1
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# Async utilities
asyncio-throttle==1.0.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==41.0.8

# Database (for future scaling)
sqlalchemy==2.0.23
alembic==1.13.1

# Additional utilities
python-dateutil==2.8.2
